requirements.txt
# AI Providers (at least one is required)
google-generativeai>=0.3.0
openai>=1.3.0
anthropic>=0.8.0

# Core dependencies
python-dotenv>=1.0.0
gitpython>=3.1.40
click>=8.1.7
rich>=13.6.0
pydantic>=2.5.0
pytest>=7.4.3
markdown>=3.5.0
